requirements.txt
# --- Core ---
Django>=1.2.4,<1.3

# --- Common ---
django-staticfiles==0.3.4
